+++ src/dog_marker/database/cruds/entry_crud.py
@@ -1,75 +1,209 @@
-from datetime import datetime
+import datetime
+from operator import and_
+from typing import Callable, Type, Iterable
 from uuid import UUID
 
-from typing import Iterable, Protocol
-
-from sqlalchemy import and_, desc
+from result import Result, Err, Ok
+from sqlalchemy import desc
 from sqlalchemy.orm import Session, Query
+from sqlalchemy.sql.functions import count
 
-from dog_marker.dtypes.coordinate import Coordinate, Longitude, Latitude
 from ..errors import DbNotFoundError
-from ..models import EntryDbModel, HiddenEntry, EntryImageDbModel, CategoryDbModel
-from dog_marker.database.schemas import Entry, WarningLevel, warning_levels
+from ..models import EntryDbModel, CategoryDbModel, EntryImageDbModel, HiddenEntry
+from ..schemas import WarningLevel, warning_levels
+from ...dtypes.coordinate import Coordinate, Longitude, Latitude
 from ...dtypes.pagination import Pagination
 
 
-class CreateEntryProtocol(Protocol):
-    id: UUID | None
-    title: str
-    longitude: Longitude
-    latitude: Latitude
-    description: str | None
-    image_path: str | None
-    warning_level: WarningLevel | warning_levels | None
-    image_delete_url: str | None
-    create_date: datetime | None
-    categories: list[str] | None
-
-
-class UpdateEntryProtocol(Protocol):
-    title: str
-    longitude: Longitude
-    latitude: Latitude
-    description: str | None
-    image_path: str | None
-    warning_level: WarningLevel | warning_levels
-    image_delete_url: str | None
-    categories: list[str]
-
-
+# noinspection PyMethodMayBeStatic
 class EntryCRUD:
     def __init__(self, db: Session):
         self.db = db
 
-    def __get(self, entry_id: UUID) -> EntryDbModel:
-        # noinspection PyTypeChecker
-        result: EntryDbModel = self.db.query(EntryDbModel).filter_by(id=entry_id, mark_to_delete=False).first()
+    def get(self, entry_id: UUID) -> Result[EntryDbModel, Exception]:
+        result: EntryDbModel = self.db.query(EntryDbModel).get(entry_id)
         if result is None:
-            raise DbNotFoundError(f"Cannot find entry with id {entry_id}")
-        return result
+            return Err(DbNotFoundError(f"Cannot find entry with id {entry_id}"))
+        return Ok(result)
 
-    def get(self, entry_id: UUID) -> Entry:
-        entry = self.__get(entry_id)
-        return entry.to_schema()
+    def query(self) -> Result[Query[Type[EntryDbModel]], Exception]:
+        return Ok(self.db.query(EntryDbModel))
 
-    def all(
-        self,
-        user_id: UUID | None = None,
-        owner_id: UUID | None = None,
-        page_info: Pagination | None = None,
-        coordinate: Coordinate | None = None,
-        date_from: datetime | None = None,
-        warning_level: WarningLevel | warning_levels | None = None,
-        deleted: bool | None = None,
-    ) -> Iterable[Entry]:
-        # noinspection PyTypeChecker
-        query: Query[EntryDbModel] = self.db.query(EntryDbModel)
-        query = query.filter_by(mark_to_delete=deleted or False)
-
-        if owner_id is not None:
+    def all(self, page_info: Pagination) -> Callable[[Query[Type[EntryDbModel]]], Iterable[EntryDbModel]]:
+        def __internal(query: Query[Type[EntryDbModel]]) -> Iterable[EntryDbModel]:
+            if page_info:
+                query = query.offset(page_info.skip).limit(page_info.limit)
             # noinspection PyTypeChecker
-            query = query.filter(EntryDbModel.user_id == owner_id)
-        elif user_id is not None:
+            return query.all()
+
+        return __internal
+
+    def add(self) -> Callable[[EntryDbModel], EntryDbModel]:
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            self.db.add(entry)
+            return entry
+
+        return __internal
+
+    def delete(self, user_id: UUID) -> Callable[[EntryDbModel], EntryDbModel]:
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            hidden_entry = self.db.query(HiddenEntry).filter_by(entry_id=entry.id, user_id=user_id).first()
+            if not hidden_entry:
+                hidden_entry = HiddenEntry(entry_id=entry.id, user_id=user_id)
+                self.db.add(hidden_entry)
+            return entry
+
+        return __internal
+
+    def commit(self) -> Callable[[EntryDbModel], EntryDbModel]:
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            self.db.commit()
+            return entry
+
+        return __internal
+
+    def create(self, owner_id: UUID, title: str) -> Result[EntryDbModel, Exception]:
+        model = EntryDbModel(user_id=owner_id, title=title)
+        return Ok(model)
+
+    def set_title(self, title: str):
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            entry.title = title
+            return entry
+
+        return __internal
+
+    def set_description(self, description: str | None = None) -> Callable[[EntryDbModel], EntryDbModel]:
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            entry.description = description
+            return entry
+
+        return __internal
+
+    def set_warning_level(
+        self, warning_level: WarningLevel | warning_levels | None = None
+    ) -> Callable[[EntryDbModel], EntryDbModel]:
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            entry.warning_level = WarningLevel.from_(warning_level)
+            return entry
+
+        return __internal
+
+    def set_coordinate(self, longitude: Longitude, latitude: Latitude) -> Callable[[EntryDbModel], EntryDbModel]:
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            entry.longitude = longitude
+            entry.latitude = latitude
+            return entry
+
+        return __internal
+
+    def set_categories(self, categories: list[str] | None) -> Callable[[EntryDbModel], Result[EntryDbModel, Exception]]:
+        def __internal(entry: EntryDbModel) -> Result[EntryDbModel, Exception]:
+            entry.clear_categories()
+            if not categories:
+                return Ok(entry)
+            for category_key in categories:
+                category = self.db.query(CategoryDbModel).get(category_key)
+                if not category:
+                    return Err(DbNotFoundError(f"Cannot find category with id {category_key}"))
+
+                entry.append_category(category)
+            return Ok(entry)
+
+        return __internal
+
+    def add_image(self, image_path: str | None, image_delete_url: str | None) -> Callable[[EntryDbModel], EntryDbModel]:
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            if entry.image_delete_url != image_delete_url or entry.image_path != image_path:
+                new_entry_image = EntryImageDbModel(
+                    entry_id=entry.id,
+                    image_path=image_path,
+                    image_delete_url=image_delete_url,
+                )
+                # Todo: No Add
+                self.db.add(new_entry_image)
+            return entry
+
+        return __internal
+
+    def set_id(self, entry_id: UUID | None) -> Callable[[EntryDbModel], EntryDbModel]:
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            if entry_id:
+                entry.id = entry_id
+            return entry
+
+        return __internal
+
+    def set_create_date(self, create_date: datetime.datetime | None) -> Callable[[EntryDbModel], EntryDbModel]:
+        def __internal(entry: EntryDbModel) -> EntryDbModel:
+            entry.create_date = create_date
+            return entry
+
+        return __internal
+
+    def order_by_coordinate(
+        self,
+        coordinate: Coordinate | None = None,
+    ) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+        def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
+            if coordinate:
+                query = query.order_by(desc(EntryDbModel.calc_distance(coordinate.longitude, coordinate.latitude)))
+            return query
+
+        return __internal
+
+    def filter_by_date_from(
+        self, date_from: datetime.datetime | None = None
+    ) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+        def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
+            if date_from is not None:
+                query = query.filter(EntryDbModel.update_date >= date_from)
+            return query
+
+        return __internal
+
+    def filter_by_warning_level(
+        self, warning_level: WarningLevel | warning_levels | None = None
+    ) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+        def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
+            level_enum = WarningLevel.from_(warning_level)
+            query = query.filter(EntryDbModel.warning_level >= level_enum.value)
+            return query
+
+        return __internal
+
+    def filter_by_user(self, user_id: UUID) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+        def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
+            # noinspection PyTypeChecker
+            query = query.filter(EntryDbModel.user_id == user_id)
+            return query
+
+        return __internal
+
+    def filter_deleted(
+        self, user_id: UUID | None = None
+    ) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+        def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
+            user_clause = HiddenEntry.user_id == EntryDbModel.user_id
+            if user_id:
+                user_clause |= HiddenEntry.user_id == user_id
+
+            query = query.join(
+                HiddenEntry,
+                onclause=and_(
+                    HiddenEntry.entry_id == EntryDbModel.id,
+                    user_clause,
+                ),
+                isouter=True,
+            )
+            # noinspection PyTypeChecker
+            query = query.filter(HiddenEntry.entry_id == None)  # noqa: E711
+            return query
+
+        return __internal
+
+    def filter_show_deleted(self, user_id: UUID) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+        def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
             query = query.join(
                 HiddenEntry,
                 onclause=and_(
@@ -79,99 +213,19 @@ class EntryCRUD:
                 isouter=True,
             )
             # noinspection PyTypeChecker
-            query = query.filter(HiddenEntry.entry_id == None)  # noqa: E711
+            query = query.filter(HiddenEntry.entry_id != None)  # noqa: E711
+            return query
 
-        if coordinate:
-            query = query.order_by(desc(EntryDbModel.calc_distance(coordinate.longitude, coordinate.latitude)))
+        return __internal
 
-        if date_from is not None:
-            query = query.filter(EntryDbModel.update_date >= date_from)
-
-        if warning_level:
-            level_enum = WarningLevel.from_(warning_level)
-            query = query.filter(EntryDbModel.warning_level >= level_enum.value)
-
-        if page_info is not None:
-            query = query.offset(page_info.skip).limit(page_info.limit)
-
-        for entry in query.all():
-            yield entry.to_schema()
-
-    def create(self, user_id: UUID, data: CreateEntryProtocol) -> Entry:
-        new_entry = EntryDbModel(
-            id=data.id,
-            user_id=user_id,
-            title=data.title,
-            description=data.description,
-            warning_level=WarningLevel.from_(data.warning_level),
-            longitude=data.longitude,
-            latitude=data.latitude,
-            create_date=data.create_date,
-        )
-
-        if data.image_path or data.image_delete_url:
-            new_entry.image_info = EntryImageDbModel(
-                image_path=data.image_path,
-                image_delete_url=data.image_delete_url,
+    def filter_owner_deleted(self) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+        def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
+            sub_query = (
+                self.db.query(count(HiddenEntry.user_id).label("count"))
+                .filter(HiddenEntry.user_id == EntryDbModel.user_id)
+                .subquery()
             )
+            query = query.filter(sub_query == 0)
+            return query
 
-        if data.categories:
-            for category_key in data.categories:
-                category = self.db.query(CategoryDbModel).get(category_key)
-                if not category:
-                    raise DbNotFoundError(f"Cannot find category with id {category_key}")
-
-                new_entry.append_category(category)
-
-        self.db.add(new_entry)
-        self.db.commit()
-
-        return new_entry.to_schema()
-
-    def update(self, entry_id: UUID, data: UpdateEntryProtocol) -> Entry:
-        entry = self.__get(entry_id)
-
-        if entry.image_delete_url != data.image_delete_url or entry.image_path != data.image_path:
-            new_entry_image = EntryImageDbModel(
-                entry_id=entry.id,
-                image_path=data.image_path,
-                image_delete_url=data.image_delete_url,
-            )
-            self.db.add(new_entry_image)
-
-        entry.title = data.title
-        entry.warning_level = WarningLevel.from_(data.warning_level)
-        entry.longitude = data.longitude
-        entry.latitude = data.latitude
-        entry.description = data.description
-
-        entry.clear_categories()
-        for category_key in data.categories:
-            category = self.db.query(CategoryDbModel).get(category_key)
-            if not category:
-                raise DbNotFoundError(f"Cannot find category with id {category_key}")
-
-            entry.append_category(category)
-
-        self.db.commit()
-
-        return entry.to_schema()
-
-    def delete(self, entry_id: UUID, user_id: UUID) -> bool:
-        entry = self.__get(entry_id)
-
-        if entry.user_id == user_id:
-            entry.mark_to_delete = True
-            self.db.commit()
-            return True
-
-        hidden_entry = self.db.query(HiddenEntry).filter_by(entry_id=entry.id, user_id=user_id).first()
-
-        if hidden_entry:
-            return True
-
-        hidden_entry = HiddenEntry(entry_id=entry.id, user_id=user_id)
-        self.db.add(hidden_entry)
-        self.db.commit()
-
-        return True
+        return __internal